Rearranges db-accessor test suites so that fully mocked worker tests - those testing start-up and clustering behaviour, do not use the dbBaseSuite.
The tests for DB namespaces now have their own suite.

The ensureStartup method had extraneous behaviour that was removed,
slightly affecting the namespace tests.

Various relocations of methods between child and base suites are
included in aid of the separation.
